sched: fixes sched_boost_disable
Everytime a boost level is disabled, if its refcount is zero, its exit
routine and the entry routine of the next boost level are called. This
causes an issue when the previous boost level and the next boost level
are same as it increments the boost count of the next boost level.This
is fixed by not calling the entry and exit routines, when the previous
and the next boost levels are same.

Change-Id: I482d0ee70c082bc647768958ffa5fcd6f5ee8e5b
Signed-off-by: Ankit Sharma <quic_anshar@quicinc.com>